changing to compute fibonacci - more interesting
