+++ autofac/Autofac101/Autofac101/Services/ComputeFibonacci/IComputeFibonacci.cs
@@ -1,0 +1,7 @@
+﻿namespace Autofac101.Services.ComputeFibonacci
+{
+	public interface IComputeSquares
+	{
+		void Compute();
+	}
+}
