--- autofac/Autofac101/Autofac101/Services/ComputeSquares/IComputeSquares.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Autofac101.Services.ComputeSquares
-{
-	public interface IComputeSquares
-	{
-		void Compute();
-	}
-}
